// src/app/core/api/generated/model/attendanceRawDto.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PunchDirection } from './punchDirection';
import { AttendanceSource } from './attendanceSource';


export interface AttendanceRawDto { 
    id?: string;
    empId?: string;
    deviceId?: string | null;
    fingerTime?: string;
    direction?: PunchDirection;
    source?: AttendanceSource;
    isProcessed?: boolean;
}